--- a/main/src/ca/uwaterloo/flix/language/phase/jvm/ClassMaker.scala
+++ b/main/src/ca/uwaterloo/flix/language/phase/jvm/ClassMaker.scala
@@ -244,16 +244,38 @@ object ClassMaker {
   }
 
   sealed case class StaticField(clazz: JvmName, name: String, tpe: BackendType) {
-    def mkField(cm: InstanceClassMaker, v: Visibility, f: Final): Unit =
+    def mkStaticField(cm: InstanceClassMaker, v: Visibility, f: Final): Unit =
       cm.mkStaticField(name, tpe, v, f)
 
-    def mkField(cm: AbstractClassMaker, v: Visibility, f: Final): Unit =
+    def mkStaticField(cm: AbstractClassMaker, v: Visibility, f: Final): Unit =
       cm.mkStaticField(name, tpe, v, f)
 
-    def putField(): InstructionSet =
+    def putStaticField(): InstructionSet =
       PUTSTATIC(clazz, name, tpe)
 
-    def getField(): InstructionSet =
+    def getStaticField(): InstructionSet =
       GETSTATIC(clazz, name, tpe)
+  }
+
+  sealed case class InstanceMethod(clazz: JvmName, name: String, d: MethodDescriptor) {
+    def mkMethod(cm: InstanceClassMaker, ins: InstructionSet, v: Visibility, f: Final): Unit =
+      cm.mkMethod(ins, name, d, v, f)
+
+    def mkMethod(cm: AbstractClassMaker, ins: InstructionSet, v: Visibility, f: Final): Unit =
+      cm.mkMethod(ins, name, d, v, f)
+
+    def invoke(): InstructionSet =
+      INVOKEVIRTUAL(clazz, name, d)
+  }
+
+  sealed case class StaticMethod(clazz: JvmName, name: String, d: MethodDescriptor) {
+    def mkStaticMethod(cm: InstanceClassMaker, ins: InstructionSet, v: Visibility, f: Final): Unit =
+      cm.mkStaticMethod(ins, name, d, v, f)
+
+    def mkStaticMethod(cm: AbstractClassMaker, ins: InstructionSet, v: Visibility, f: Final): Unit =
+      cm.mkStaticMethod(ins, name, d, v, f)
+
+    def invokeStatic(): InstructionSet =
+      INVOKESTATIC(clazz, name, d)
   }
 }