--- a/main/src/ca/uwaterloo/flix/language/phase/jvm/ClassMaker.scala
+++ b/main/src/ca/uwaterloo/flix/language/phase/jvm/ClassMaker.scala
@@ -264,17 +264,23 @@ object ClassMaker {
     def mkMethod(cm: AbstractClassMaker, ins: InstructionSet, v: Visibility, f: Final): Unit =
       cm.mkMethod(ins, name, d, v, f)
 
-    def invoke(): InstructionSet =
-      INVOKEVIRTUAL(clazz, name, d)
-  }
-
-  sealed case class AbstractMethod(clazz: JvmName, name: String, d: MethodDescriptor) {
-    def implementation(clazz: JvmName): InstanceMethod = InstanceMethod(clazz, name, d)
-
     def mkAbstractMethod(cm: AbstractClassMaker): Unit =
       cm.mkAbstractMethod(name, d)
 
     def mkAbstractMethod(cm: InterfaceMaker): Unit =
+      cm.mkAbstractMethod(name, d)
+
+    def invoke(): InstructionSet =
+      INVOKEVIRTUAL(clazz, name, d)
+  }
+
+  sealed case class InterfaceMethod(clazz: JvmName, name: String, d: MethodDescriptor) {
+    def implementation(clazz: JvmName): InstanceMethod = InstanceMethod(clazz, name, d)
+
+    def mkInterfaceMethod(cm: AbstractClassMaker): Unit =
+      cm.mkAbstractMethod(name, d)
+
+    def mkInterfaceMethod(cm: InterfaceMaker): Unit =
       cm.mkAbstractMethod(name, d)
 
     def invokeInterface(): InstructionSet =