--- a/main/src/ca/uwaterloo/flix/language/phase/jvm/GenFunctionClasses.scala
+++ b/main/src/ca/uwaterloo/flix/language/phase/jvm/GenFunctionClasses.scala
@@ -115,7 +115,7 @@ object GenFunctionClasses {
     val backendContinuationType = BackendObjType.Continuation(BackendType.toErasedBackendType(closureResultType))
 
     // Method header
-    val m = visitor.visitMethod(ACC_PUBLIC + ACC_FINAL, backendContinuationType.InvokeMethodName,
+    val m = visitor.visitMethod(ACC_PUBLIC + ACC_FINAL, backendContinuationType.InvokeMethod.name,
       AsmOps.getMethodDescriptor(Nil, continuationType), null, null)
 
     // Enter label