--- a/main/src/ca/uwaterloo/flix/language/phase/jvm/GenGlobalCounterClass.scala
+++ b/main/src/ca/uwaterloo/flix/language/phase/jvm/GenGlobalCounterClass.scala
@@ -41,7 +41,7 @@ object GenGlobalCounterClass {
 
     cm.mkObjectConstructor(IsPrivate)
     cm.mkStaticConstructor(genStaticConstructor())
-    counterField.mkField(cm, IsPublic, IsFinal)
+    counterField.mkStaticField(cm, IsPublic, IsFinal)
     cm.mkStaticMethod(genNewIdMethod(), NewIdMethodName, MethodDescriptor(Nil, BackendType.Int64), IsPublic, IsFinal)
 
     cm.closeClassMaker
@@ -50,11 +50,11 @@ object GenGlobalCounterClass {
   private def genStaticConstructor(): InstructionSet =
     NEW(JvmName.AtomicLong) ~
       DUP() ~ invokeConstructor(JvmName.AtomicLong) ~
-      counterField.putField() ~
+      counterField.putStaticField() ~
       RETURN()
 
   private def genNewIdMethod()(implicit flix: Flix): InstructionSet =
-    counterField.getField() ~
+    counterField.getStaticField() ~
       INVOKEVIRTUAL(JvmName.AtomicLong, "getAndIncrement", MethodDescriptor(Nil, BackendType.Int64)) ~
       LRETURN()