--- a/main/src/ca/uwaterloo/flix/language/phase/jvm/GenNamespaceClasses.scala
+++ b/main/src/ca/uwaterloo/flix/language/phase/jvm/GenNamespaceClasses.scala
@@ -111,7 +111,7 @@ object GenNamespaceClasses {
       // Incrementing the offset
       offset += AsmOps.getStackSize(arg)
     }
-    method.visitMethodInsn(INVOKEVIRTUAL, functionType.name.toInternalName, backendContinuationType.UnwindMethodName, AsmOps.getMethodDescriptor(Nil, erasedResult), false)
+    method.visitMethodInsn(INVOKEVIRTUAL, functionType.name.toInternalName, backendContinuationType.UnwindMethod.name, AsmOps.getMethodDescriptor(Nil, erasedResult), false)
     // no erasure here because the ns function works on erased values
 
     // Return