--- a/main/src/ca/uwaterloo/flix/language/phase/jvm/JvmBackend.scala
+++ b/main/src/ca/uwaterloo/flix/language/phase/jvm/JvmBackend.scala
@@ -121,7 +121,7 @@ object JvmBackend extends Phase[Root, CompilationResult] {
 
       //
       // Generate record interface.
-      // TODO
+      //
       val recordInterfaces = GenRecordInterface.gen()
 
       //