Now without nullpointer exceptions. Related to #2591.