@magnus-madsen 2021-12-28
Does this have any performance impact?