@JonathanStarup 2021-12-29
<details>
<summary>Benchmarks</summary>

Phase | PR | Master
:--- | ---: | ---:
Deriver | 0,8 ms (00,0%) | 0,8 ms (00,0%)
Reader | 1,2 ms (00,0%) | 1,2 ms (00,0%)
Instances | 2,0 ms (00,1%) | 2,2 ms (00,1%)
Terminator | 9,4 ms (00,3%) | 8,2 ms (00,2%)
Tailrec | 9,7 ms (00,3%) | 9,7 ms (00,3%)
Safety | 13,2 ms (00,4%) | 10,7 ms (00,3%)
Lowering | 15,0 ms (00,4%) | 13,4 ms (00,4%)
Stratifier | 19,2 ms (00,5%) | 14,8 ms (00,4%)
Simplifier | 19,5 ms (00,5%) | 18,4 ms (00,5%)
VarNumbering | 19,9 ms (00,6%) | 20,2 ms (00,6%)
PatternExhaustiveness | 20,6 ms (00,6%) | 20,2 ms (00,6%)
Optimizer | 22,3 ms (00,6%) | 22,1 ms (00,6%)
Kinder | 22,9 ms (00,6%) | 22,2 ms (00,6%)
Eraser | 25,6 ms (00,7%) | 26,1 ms (00,7%)
LambdaLift | 27,2 ms (00,8%) | 27,0 ms (00,8%)
ClosureConv | 52,7 ms (01,5%) | 53,8 ms (01,5%)
Weeder | 55,1 ms (01,5%) | 54,2 ms (01,5%)
Redundancy | 69,4 ms (01,9%) | 57,6 ms (01,6%)
Namer | 70,5 ms (02,0%) | 69,1 ms (02,0%)
TreeShaker | 75,6 ms (02,1%) | 73,9 ms (02,1%)
Parser | 97,4 ms (02,7%) | 88,0 ms (02,5%)
Finalize | 103,0 ms (02,9%) | 110,1 ms (03,1%)
Resolver | 197,9 ms (05,6%) | 189,7 ms (05,4%)
Monomorph | 443,2 ms (12,4%) | 450,9 ms (12,7%)
Typer | 837,4 ms (23,5%) | 864,2 ms (24,4%)
JvmBackend | 1330,3 ms (37,4%) | 1310,0 ms (37,0%)

Throughput | PR | Master
:--- | ---: | ---:
min | 5.963 | 6.083
max | 16.332 | 16.452
avg | 14.669 | 14.669
median | 16.103 | 15.755
</details>